Moving full project

+++ Gwent/Models/Card.cs
@@ -2,7 +2,7 @@
 
 public class Card
 {
-    public Card(int id, Role role, int basePower, GameImpact ownImpact)
+    public Card(byte id, Role role, int basePower, GameImpact ownImpact)
     {
         Id = id;
         Role = role;
@@ -10,7 +10,7 @@ public class Card
         OwnImpact = ownImpact;
     }
 
-    public int Id { get; }
+    public byte Id { get; }
     public Role Role { get; }
     private int BasePower { get; }
